Add mbedtls_nv_seed support
Add support for mbed TLS Non-Volatile seed as an entropy source by
using the config store. The patch implements two functions in
mbedtls/platform/src/mbed_nv_seed.cpp.

  * mbedtls_nv_seed_read(): called by mbed TLS whenever the entropy
    source is polled. This reads the stored NV seed and copies it into
    a buffer.
  * mbedtls_nv_seed_write(): called by mbed TLS only once per
    initialised context. This updates the NV seed with the supplied
    value.

Because of the nature of the mbed TLS API, this uses the synchronous
C++ wrapper API around cfstore. The functionality is enabled by default
in targets where the config store with Non-Volatile storage is enabled.
However, it can be disabled by the user by undefining the macro
MBEDTLS_ENTROPY_NV_SEED through the user defined mbed TLS config.h.

WARNING!!! The config store implements some wear-leveling features to
avoid early wear-out of the flash. This means that when the NV seed is
updated, previous seed values will potentially be left in the flash
unprotected. Ideally, we would have functionality in the config store
that triggers an erase operation on the previously used blocks, but
this is currently not available.